% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exppackage-package.R
\docType{package}
\name{exppackage}
\alias{exppackage}
\title{exppackage: An example package to demonstrate usethis}
\description{
bLABLABLABLABLABLABLABLAblA
}
\section{Section 1}{

blablablba
}

\section{Section 2}{

yadeeyadeeya
}

