% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_name.R
\name{linear_transform}
\alias{linear_transform}
\title{Adds 1 to a number}
\usage{
linear_transform(x)
}
\arguments{
\item{x}{is a number}
}
\value{
x+1
}
\description{
Adds 1 to a number
}
\examples{
linear_transform(1)
}
